styles(web): place form into its own file
